requirements.txt
Django== 2.2.0
django-bootstrap4 ==1.1.1
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
django-redis==4.10.0
django-user-agents==0.3.2
django-webpack-loader==0.6.0
djangorestframework==3.8.2
dnspython==1.16.0
django_heroku == 0.3.1
gunicorn==19.0.0
whitenoise == 5.0.1
channels == 2.4.0
opencv-python-headless == 4.1.0.25
six == 1.12.0